Post-churn simplification: dedupe export, batch helpers, verb factory (0.5.3)
Consolidates the cruft that accumulated across PRs #41-#65 (net -204
lines, zero behavior change except one fix):
- task export now shares bulk export-tasks' implementation — fixing its
  CSV missing the date_created/date_updated/url columns
- batch result partitioning/failure reporting extracted from three
  copy-pasted sites; --ids-only/--brief guard and the 11x duplicated
  --brief help string live in shared.py
- done/close/start/park generated from a status-verb table (82 -> ~30
  lines, per-command help verified)
- bulk.py error chains standardized on handle_clickup_errors with a
  targeted OSError catch around file writes
- PR-named test files dissolved back into feature-named files (the
  _plain_help ANSI normalizer moved to conftest); AGENT.md spinner-shim
  text updated to reflect their removal
