+++ clickup/__init__.py
@@ -1,6 +1,6 @@
 """ClickUp Toolkit - A powerful CLI for ClickUp task management."""
 
-__version__ = "0.5.2"
+__version__ = "0.5.3"
 
 # Import main modules
 from . import cli, core
