+++ clickup/cli/__init__.py
@@ -1,6 +1,6 @@
 """ClickUp Toolkit CLI - Command-line interface for ClickUp."""
 
-__version__ = "0.5.2"
+__version__ = "0.5.3"
 
 from .main import app, main
