+++ tests/integration/test_task_commands.py
@@ -19,7 +19,7 @@ from clickup.core.exceptions import ClickUpError
 from clickup.core.models import Comment, PriorityInfo, StatusInfo, Task, Team, User
 from clickup.core.models import List as ClickUpList
 
-from .conftest import make_mock_ctx
+from .conftest import _strip_ansi, make_mock_ctx
 
 runner = CliRunner()
 
@@ -1409,7 +1409,9 @@ def test_task_export_csv_cov(mock_get_client):
     from pathlib import Path
 
     content = Path(f.name).read_text()
-    assert "id,name,status" in content
+    # Canonical CSV columns (shared with bulk export-tasks): includes
+    # date_created, date_updated, url that the old task export was missing.
+    assert "id,name,description,status,priority,assignees,due_date,date_created,date_updated,url" in content
     assert "One" in content
 
 
@@ -1630,3 +1632,739 @@ def test_task_status_brief_omits_description(mock_get_client):
     data = json.loads(result.stdout)
     assert data["id"] == "t1"
     assert "description" not in data
+
+
+# =============================================================================
+# Batch task create (from backlog-49 item 1)
+# =============================================================================
+
+
+def _make_task_b49(id: str, name: str, **overrides) -> Task:
+    """Build a real Task model for tests that need model_dump."""
+    kwargs = {"id": id, "name": name}
+    kwargs.update(overrides)
+    return Task(**kwargs)
+
+
+@patch("clickup.cli.commands.task.get_client")
+def test_batch_create_three_tasks_collection_envelope(mock_get_client):
+    """3 names -> 3 tasks in order, collection envelope, flags applied to all."""
+    mock_client = AsyncMock()
+    mock_client.create_task.side_effect = [
+        _make_task_b49("t1", "Alpha", url="https://x/t1"),
+        _make_task_b49("t2", "Bravo", url="https://x/t2"),
+        _make_task_b49("t3", "Charlie", url="https://x/t3"),
+    ]
+    mock_get_client.return_value = make_mock_ctx(mock_client)
+
+    result = runner.invoke(
+        app,
+        [
+            "task",
+            "create",
+            "Alpha",
+            "Bravo",
+            "Charlie",
+            "--list-id",
+            "L1",
+            "--description",
+            "shared desc",
+            "--priority",
+            "2",
+        ],
+    )
+
+    assert result.exit_code == 0
+    data = json.loads(result.stdout)
+    assert data["count"] == 3
+    assert [t["name"] for t in data["data"]] == ["Alpha", "Bravo", "Charlie"]
+
+    # Verify flags applied to every call
+    for call in mock_client.create_task.await_args_list:
+        assert call.kwargs["description"] == "shared desc"
+        assert call.kwargs["priority"] == 2
+
+
+@patch("clickup.cli.commands.task.get_client")
+def test_single_create_unchanged_shape(mock_get_client):
+    """Single name produces a singleton task object, NOT a collection envelope."""
+    mock_client = AsyncMock()
+    mock_client.create_task.return_value = _make_task_b49("t1", "Solo", url="https://x/t1")
+    mock_get_client.return_value = make_mock_ctx(mock_client)
+
+    result = runner.invoke(app, ["task", "create", "Solo", "--list-id", "L1"])
+
+    assert result.exit_code == 0
+    data = json.loads(result.stdout)
+    assert data["id"] == "t1"
+    assert data["name"] == "Solo"
+    # No "data" / "count" keys — singleton shape
+    assert "data" not in data
+    assert "count" not in data
+
+
+@patch("clickup.cli.commands.task.get_client")
+def test_batch_create_partial_failure_exit_1(mock_get_client):
+    """When some creates fail: successes rendered, failures warned, exit 1."""
+    mock_client = AsyncMock()
+    mock_client.create_task.side_effect = [
+        _make_task_b49("t1", "Good"),
+        ClickUpError("rate limited"),
+        _make_task_b49("t3", "Also Good"),
+    ]
+    mock_get_client.return_value = make_mock_ctx(mock_client)
+
+    result = runner.invoke(
+        app,
+        ["task", "create", "Good", "Bad", "Also Good", "--list-id", "L1"],
+    )
+
+    assert result.exit_code == 1
+    data = json.loads(result.stdout)
+    # Only the 2 successes appear in the envelope
+    assert data["count"] == 2
+    assert [t["name"] for t in data["data"]] == ["Good", "Also Good"]
+    # Error on stderr
+    assert "Failed to create task 'Bad'" in result.stderr
+
+
+@patch("clickup.cli.commands.task.get_client")
+def test_batch_create_brief(mock_get_client):
+    """--brief works for batch creates."""
+    mock_client = AsyncMock()
+    mock_client.create_task.side_effect = [
+        _make_task_b49("t1", "A", description="long text", url="https://x"),
+        _make_task_b49("t2", "B", description="long text 2", url="https://x"),
+    ]
+    mock_get_client.return_value = make_mock_ctx(mock_client)
+
+    result = runner.invoke(
+        app,
+        ["task", "create", "A", "B", "--list-id", "L1", "--brief"],
+    )
+
+    assert result.exit_code == 0
+    data = json.loads(result.stdout)
+    assert data["count"] == 2
+    for t in data["data"]:
+        assert "description" not in t
+
+
+def test_batch_create_validates_each_name():
+    """Each name is validated — empty names rejected."""
+    result = runner.invoke(app, ["task", "create", "Good", "   ", "--list-id", "L1"])
+    assert result.exit_code == 2
+    assert "empty" in result.stderr.lower()
+
+
+# =============================================================================
+# Zero-result hints (from backlog-49 item 3)
+# =============================================================================
+
+
+@patch("clickup.cli.commands.task.get_client")
+def test_zero_result_hint_filtered_empty_table_mode(mock_get_client):
+    """Filtered-empty emits info message with pre-filter count in table mode."""
+    mock_client = AsyncMock()
+    tasks = [
+        _make_task_b49(f"t{i}", f"Task {i}", status=StatusInfo(status="complete", type="closed")) for i in range(5)
+    ]
+    mock_client.get_tasks.return_value = tasks
+    mock_get_client.return_value = make_mock_ctx(mock_client)
+
+    result = runner.invoke(
+        app,
+        ["--format", "table", "task", "list", "--list-id", "L1", "--open-only"],
+    )
+
+    assert result.exit_code == 0
+    combined = _strip_ansi(result.stdout + result.stderr)
+    assert "0 tasks matched the active filters" in combined
+    assert "5 tasks total" in combined
+
+
+@patch("clickup.cli.commands.task.get_client")
+def test_zero_result_hint_json_mode_suppressed(mock_get_client):
+    """In JSON mode, info-level hint is suppressed; only envelope on stdout."""
+    mock_client = AsyncMock()
+    tasks = [
+        _make_task_b49(f"t{i}", f"Task {i}", status=StatusInfo(status="complete", type="closed")) for i in range(3)
+    ]
+    mock_client.get_tasks.return_value = tasks
+    mock_get_client.return_value = make_mock_ctx(mock_client)
+
+    result = runner.invoke(
+        app,
+        ["task", "list", "--list-id", "L1", "--open-only"],
+    )
+
+    assert result.exit_code == 0
+    data = json.loads(result.stdout)
+    assert data == {"data": [], "count": 0}
+    assert "0 tasks matched" not in result.stdout
+
+
+@patch("clickup.cli.commands.task.get_client")
+def test_zero_result_no_hint_when_unfiltered(mock_get_client):
+    """Unfiltered empty list emits 'No tasks found' — no hint about filters."""
+    mock_client = AsyncMock()
+    mock_client.get_tasks.return_value = []
+    mock_get_client.return_value = make_mock_ctx(mock_client)
+
+    result = runner.invoke(
+        app,
+        ["--format", "table", "task", "list", "--list-id", "L1"],
+    )
+
+    assert result.exit_code == 0
+    combined = _strip_ansi(result.stdout + result.stderr)
+    assert "0 tasks matched the active filters" not in combined
+
+
+@patch("clickup.cli.commands.task.get_client")
+def test_zero_result_hint_task_search_with_filters(mock_get_client):
+    """task search emits zero-result hint when filters are active."""
+    mock_client = AsyncMock()
+    mock_client.search_tasks.return_value = [
+        _make_task_b49("t1", "Alpha", status=StatusInfo(status="open")),
+    ]
+    mock_get_client.return_value = make_mock_ctx(mock_client)
+
+    result = runner.invoke(
+        app,
+        ["--format", "table", "task", "search", "--query", "alpha", "--status", "closed", "--workspace-id", "W1"],
+    )
+
+    assert result.exit_code == 0
+    combined = _strip_ansi(result.stdout + result.stderr)
+    assert "0 tasks matched the active filters" in combined
+
+
+@patch("clickup.cli.commands.task.get_client")
+def test_zero_result_hint_task_mine_with_filters(mock_get_client):
+    """task mine emits zero-result hint when filters are active."""
+    mock_client = AsyncMock()
+    mock_client.get_user.return_value = User(id=42, username="evan", email="e@x.com")
+    mock_client.search_tasks.return_value = [
+        _make_task_b49("t1", "My Task", status=StatusInfo(status="open")),
+    ]
+    mock_get_client.return_value = make_mock_ctx(mock_client)
+
+    result = runner.invoke(
+        app,
+        ["--format", "table", "task", "mine", "--workspace-id", "W1", "--open-only"],
+    )
+
+    mock_client.search_tasks.return_value = [
+        _make_task_b49("t1", "My Task", status=StatusInfo(status="open")),
+    ]
+    result = runner.invoke(
+        app,
+        ["--format", "table", "task", "mine", "--workspace-id", "W1", "--status", "nonexistent"],
+    )
+
+    assert result.exit_code == 0
+    combined = _strip_ansi(result.stdout + result.stderr)
+    assert "0 tasks matched the active filters" in combined
+
+
+# =============================================================================
+# task search --name-only (from backlog-49 item 4)
+# =============================================================================
+
+
+@patch("clickup.cli.commands.task.get_client")
+def test_name_only_filters_by_task_name(mock_get_client):
+    """--name-only keeps only tasks whose name contains the query (case-insensitive)."""
+    mock_client = AsyncMock()
+    mock_client.search_tasks.return_value = [
+        _make_task_b49("t1", "Deploy API"),
+        _make_task_b49("t2", "Write API docs"),
+        _make_task_b49("t3", "Fix login bug"),
+    ]
+    mock_get_client.return_value = make_mock_ctx(mock_client)
+
+    result = runner.invoke(
+        app,
+        ["task", "search", "--query", "api", "--workspace-id", "W1", "--name-only"],
+    )
+
+    assert result.exit_code == 0
+    data = json.loads(result.stdout)
+    assert data["count"] == 2
+    names = [t["name"] for t in data["data"]]
+    assert "Deploy API" in names
+    assert "Write API docs" in names
+    assert "Fix login bug" not in names
+
+
+@patch("clickup.cli.commands.task.get_client")
+def test_name_only_case_insensitive(mock_get_client):
+    """--name-only matching is case-insensitive."""
+    mock_client = AsyncMock()
+    mock_client.search_tasks.return_value = [
+        _make_task_b49("t1", "URGENT Bug Fix"),
+        _make_task_b49("t2", "urgent refactor"),
+        _make_task_b49("t3", "Normal task"),
+    ]
+    mock_get_client.return_value = make_mock_ctx(mock_client)
+
+    result = runner.invoke(
+        app,
+        ["task", "search", "--query", "urgent", "--workspace-id", "W1", "--name-only"],
+    )
+
+    assert result.exit_code == 0
+    data = json.loads(result.stdout)
+    assert data["count"] == 2
+
+
+@patch("clickup.cli.commands.task.get_client")
+def test_name_only_with_limit(mock_get_client):
+    """--name-only + --limit: count reflects the filtered (then truncated) set."""
+    mock_client = AsyncMock()
+    mock_client.search_tasks.return_value = [_make_task_b49(f"t{i}", f"Matching {i}") for i in range(10)]
+    mock_get_client.return_value = make_mock_ctx(mock_client)
+
+    result = runner.invoke(
+        app,
+        ["task", "search", "--query", "Matching", "--workspace-id", "W1", "--name-only", "--limit", "3"],
+    )
+
+    assert result.exit_code == 0
+    data = json.loads(result.stdout)
+    assert data["count"] == 3
+
+
+# =============================================================================
+# Due-date format consistency (from backlog-49 item 5)
+# =============================================================================
+
+
+@patch("clickup.cli.commands.task.get_client")
+def test_create_due_date_accepts_yyyy_mm_dd(mock_get_client):
+    """task create --due-date 2026-07-01 sends epoch ms to the provider."""
+    mock_client = AsyncMock()
+    mock_client.create_task.return_value = _make_task_b49("t1", "X")
+    mock_get_client.return_value = make_mock_ctx(mock_client)
+
+    result = runner.invoke(
+        app,
+        ["task", "create", "X", "--list-id", "L1", "--due-date", "2026-07-01"],
+    )
+
+    assert result.exit_code == 0
+    call_kwargs = mock_client.create_task.call_args.kwargs
+    due_val = call_kwargs["due_date"]
+    assert due_val == str(1782864000000)
+
+
+@patch("clickup.cli.commands.task.get_client")
+def test_create_due_date_accepts_relative(mock_get_client):
+    """task create --due-date 7d sends an epoch-ms in the recent past."""
+    mock_client = AsyncMock()
+    mock_client.create_task.return_value = _make_task_b49("t1", "X")
+    mock_get_client.return_value = make_mock_ctx(mock_client)
+
+    result = runner.invoke(
+        app,
+        ["task", "create", "X", "--list-id", "L1", "--due-date", "7d"],
+    )
+
+    assert result.exit_code == 0
+    call_kwargs = mock_client.create_task.call_args.kwargs
+    due_val = int(call_kwargs["due_date"])
+    assert due_val > 0
+
+
+@patch("clickup.cli.commands.task.get_client")
+def test_update_due_date_accepts_yyyy_mm_dd(mock_get_client):
+    """task update --due-date 2026-07-01 sends epoch ms to the provider."""
+    mock_client = AsyncMock()
+    mock_client.update_task.return_value = _make_task_b49("t1", "X")
+    mock_get_client.return_value = make_mock_ctx(mock_client)
+
+    result = runner.invoke(
+        app,
+        ["task", "update", "t1", "--due-date", "2026-07-01"],
+    )
+
+    assert result.exit_code == 0
+    call_kwargs = mock_client.update_task.call_args.kwargs
+    due_val = call_kwargs["due_date"]
+    assert due_val == str(1782864000000)
+
+
+@patch("clickup.cli.commands.task.get_client")
+def test_update_due_date_accepts_relative(mock_get_client):
+    """task update --due-date 7d sends epoch ms."""
+    mock_client = AsyncMock()
+    mock_client.update_task.return_value = _make_task_b49("t1", "X")
+    mock_get_client.return_value = make_mock_ctx(mock_client)
+
+    result = runner.invoke(
+        app,
+        ["task", "update", "t1", "--due-date", "7d"],
+    )
+
+    assert result.exit_code == 0
+    call_kwargs = mock_client.update_task.call_args.kwargs
+    due_val = int(call_kwargs["due_date"])
+    assert due_val > 0
+
+
+@patch("clickup.cli.commands.task.get_client")
+def test_update_due_date_accepts_epoch_ms(mock_get_client):
+    """task update --due-date <epoch-ms> passes through unchanged."""
+    mock_client = AsyncMock()
+    mock_client.update_task.return_value = _make_task_b49("t1", "X")
+    mock_get_client.return_value = make_mock_ctx(mock_client)
+
+    result = runner.invoke(
+        app,
+        ["task", "update", "t1", "--due-date", "1782691200000"],
+    )
+
+    assert result.exit_code == 0
+    call_kwargs = mock_client.update_task.call_args.kwargs
+    assert call_kwargs["due_date"] == "1782691200000"
+
+
+@patch("clickup.cli.commands.task.get_client")
+def test_create_due_date_accepts_epoch_ms(mock_get_client):
+    """task create --due-date <epoch-ms> passes through unchanged."""
+    mock_client = AsyncMock()
+    mock_client.create_task.return_value = _make_task_b49("t1", "X")
+    mock_get_client.return_value = make_mock_ctx(mock_client)
+
+    result = runner.invoke(
+        app,
+        ["task", "create", "X", "--list-id", "L1", "--due-date", "1782691200000"],
+    )
+
+    assert result.exit_code == 0
+    call_kwargs = mock_client.create_task.call_args.kwargs
+    assert call_kwargs["due_date"] == "1782691200000"
+
+
+# =============================================================================
+# --ids-only output mode (from polish-63)
+# =============================================================================
+
+
+def _make_task_p63(task_id: str, name: str = "T") -> Task:
+    return Task(id=task_id, name=name, url=f"https://x/{task_id}")
+
+
+class TestIdsOnlyCreate:
+    @patch("clickup.cli.commands.task.get_client")
+    def test_single_create(self, mock_get_client: AsyncMock) -> None:
+        mock_client = AsyncMock()
+        mock_client.create_task.return_value = _make_task_p63("abc123", "New")
+        mock_get_client.return_value = make_mock_ctx(mock_client)
+
+        result = runner.invoke(app, ["task", "create", "New", "--list-id", "L1", "--ids-only"])
+        assert result.exit_code == 0
+        assert result.stdout.strip() == "abc123"
+
+    @patch("clickup.cli.commands.task.get_client")
+    def test_batch_create(self, mock_get_client: AsyncMock) -> None:
+        mock_client = AsyncMock()
+        mock_client.create_task.side_effect = [
+            _make_task_p63("id1", "A"),
+            _make_task_p63("id2", "B"),
+        ]
+        mock_get_client.return_value = make_mock_ctx(mock_client)
+
+        result = runner.invoke(app, ["task", "create", "A", "B", "--list-id", "L1", "--ids-only"])
+        assert result.exit_code == 0
+        lines = result.stdout.strip().splitlines()
+        assert lines == ["id1", "id2"]
+
+    @patch("clickup.cli.commands.task.get_client")
+    def test_ids_only_json_mode(self, mock_get_client: AsyncMock) -> None:
+        """--ids-only prints plain IDs regardless of --format json."""
+        mock_client = AsyncMock()
+        mock_client.create_task.return_value = _make_task_p63("j1", "J")
+        mock_get_client.return_value = make_mock_ctx(mock_client)
+
+        result = runner.invoke(app, ["--format", "json", "task", "create", "J", "--list-id", "L1", "--ids-only"])
+        assert result.exit_code == 0
+        assert result.stdout.strip() == "j1"
+
+    def test_ids_only_and_brief_mutual_exclusion(self) -> None:
+        result = runner.invoke(app, ["task", "create", "X", "--list-id", "L1", "--ids-only", "--brief"])
+        assert result.exit_code == 2
+        assert "mutually exclusive" in result.stderr
+
+
+class TestIdsOnlyDone:
+    @patch("clickup.cli.commands.task.get_client")
+    def test_single_done(self, mock_get_client: AsyncMock) -> None:
+        mock_client = AsyncMock()
+        mock_client.update_task.return_value = _make_task_p63("d1", "Done")
+        mock_get_client.return_value = make_mock_ctx(mock_client)
+
+        result = runner.invoke(app, ["task", "done", "d1", "--ids-only"])
+        assert result.exit_code == 0
+        assert result.stdout.strip() == "d1"
+
+    @patch("clickup.cli.commands.task.get_client")
+    def test_batch_done(self, mock_get_client: AsyncMock) -> None:
+        mock_client = AsyncMock()
+        mock_client.update_task.side_effect = [
+            _make_task_p63("d1"),
+            _make_task_p63("d2"),
+        ]
+        mock_get_client.return_value = make_mock_ctx(mock_client)
+
+        result = runner.invoke(app, ["task", "done", "d1", "d2", "--ids-only"])
+        assert result.exit_code == 0
+        assert result.stdout.strip().splitlines() == ["d1", "d2"]
+
+    def test_mutual_exclusion(self) -> None:
+        result = runner.invoke(app, ["task", "done", "d1", "--ids-only", "--brief"])
+        assert result.exit_code == 2
+        assert "mutually exclusive" in result.stderr
+
+
+class TestIdsOnlyClose:
+    @patch("clickup.cli.commands.task.get_client")
+    def test_close_ids_only(self, mock_get_client: AsyncMock) -> None:
+        mock_client = AsyncMock()
+        mock_client.update_task.return_value = _make_task_p63("c1")
+        mock_get_client.return_value = make_mock_ctx(mock_client)
+
+        result = runner.invoke(app, ["task", "close", "c1", "--ids-only"])
+        assert result.exit_code == 0
+        assert result.stdout.strip() == "c1"
+
+
+class TestIdsOnlyStart:
+    @patch("clickup.cli.commands.task.get_client")
+    def test_start_ids_only(self, mock_get_client: AsyncMock) -> None:
+        mock_client = AsyncMock()
+        mock_client.update_task.return_value = _make_task_p63("s1")
+        mock_get_client.return_value = make_mock_ctx(mock_client)
+
+        result = runner.invoke(app, ["task", "start", "s1", "--ids-only"])
+        assert result.exit_code == 0
+        assert result.stdout.strip() == "s1"
+
+
+class TestIdsOnlyPark:
+    @patch("clickup.cli.commands.task.get_client")
+    def test_park_ids_only(self, mock_get_client: AsyncMock) -> None:
+        mock_client = AsyncMock()
+        mock_client.update_task.return_value = _make_task_p63("p1")
+        mock_get_client.return_value = make_mock_ctx(mock_client)
+
+        result = runner.invoke(app, ["task", "park", "p1", "--ids-only"])
+        assert result.exit_code == 0
+        assert result.stdout.strip() == "p1"
+
+
+class TestIdsOnlyStatus:
+    @patch("clickup.cli.commands.task.get_client")
+    def test_status_ids_only(self, mock_get_client: AsyncMock) -> None:
+        mock_client = AsyncMock()
+        mock_client.update_task.return_value = _make_task_p63("st1")
+        mock_get_client.return_value = make_mock_ctx(mock_client)
+
+        result = runner.invoke(app, ["task", "status", "st1", "complete", "--ids-only"])
+        assert result.exit_code == 0
+        assert result.stdout.strip() == "st1"
+
+    def test_mutual_exclusion(self) -> None:
+        result = runner.invoke(app, ["task", "status", "st1", "complete", "--ids-only", "--brief"])
+        assert result.exit_code == 2
+        assert "mutually exclusive" in result.stderr
+
+
+class TestIdsOnlyDelete:
+    @patch("clickup.cli.commands.task.get_client")
+    def test_single_delete(self, mock_get_client: AsyncMock) -> None:
+        mock_client = AsyncMock()
+        mock_client.delete_task.return_value = True
+        mock_get_client.return_value = make_mock_ctx(mock_client)
+
+        result = runner.invoke(app, ["task", "delete", "del1", "--force", "--ids-only"])
+        assert result.exit_code == 0
+        assert result.stdout.strip() == "del1"
+
+    @patch("clickup.cli.commands.task.get_client")
+    def test_batch_delete(self, mock_get_client: AsyncMock) -> None:
+        mock_client = AsyncMock()
+        mock_client.delete_task.return_value = True
+        mock_get_client.return_value = make_mock_ctx(mock_client)
+
+        result = runner.invoke(app, ["task", "delete", "del1", "del2", "--force", "--ids-only"])
+        assert result.exit_code == 0
+        assert result.stdout.strip().splitlines() == ["del1", "del2"]
+
+
+# =============================================================================
+# Truncated envelope flag (from polish-63)
+# =============================================================================
+
+
+class TestTruncatedFlag:
+    @patch("clickup.cli.commands.task.get_client")
+    def test_truncated_true_when_clipped(self, mock_get_client: AsyncMock) -> None:
+        """When result count > limit, envelope has truncated: true."""
+        mock_client = AsyncMock()
+        mock_client.get_tasks.return_value = [_make_task_p63(f"t{i}") for i in range(5)]
+        mock_get_client.return_value = make_mock_ctx(mock_client)
+
+        result = runner.invoke(app, ["task", "list", "--list-id", "L1", "--limit", "3"])
+        assert result.exit_code == 0
+        data = json.loads(result.stdout)
+        assert data["truncated"] is True
+        assert data["count"] == 3
+
+    @patch("clickup.cli.commands.task.get_client")
+    def test_no_truncated_when_not_clipped(self, mock_get_client: AsyncMock) -> None:
+        """When result count <= limit, truncated key is absent."""
+        mock_client = AsyncMock()
+        mock_client.get_tasks.return_value = [_make_task_p63(f"t{i}") for i in range(3)]
+        mock_get_client.return_value = make_mock_ctx(mock_client)
+
+        result = runner.invoke(app, ["task", "list", "--list-id", "L1", "--limit", "50"])
+        assert result.exit_code == 0
+        data = json.loads(result.stdout)
+        assert "truncated" not in data
+        assert data["count"] == 3
+
+    @patch("clickup.cli.commands.task.get_client")
+    def test_truncated_on_search(self, mock_get_client: AsyncMock) -> None:
+        """task search also emits truncated when clipped."""
+        mock_client = AsyncMock()
+        mock_client.search_tasks.return_value = [_make_task_p63(f"t{i}") for i in range(10)]
+        mock_get_client.return_value = make_mock_ctx(mock_client)
+
+        result = runner.invoke(app, ["task", "search", "--workspace-id", "W1", "--limit", "5"])
+        assert result.exit_code == 0
+        data = json.loads(result.stdout)
+        assert data["truncated"] is True
+        assert data["count"] == 5
+
+    @patch("clickup.cli.commands.task.get_client")
+    def test_truncated_on_mine(self, mock_get_client: AsyncMock) -> None:
+        """task mine also emits truncated when clipped."""
+        mock_client = AsyncMock()
+        mock_client.get_user.return_value = User(id=42, username="e", email="e@x.com")
+        mock_client.search_tasks.return_value = [_make_task_p63(f"t{i}") for i in range(5)]
+        mock_get_client.return_value = make_mock_ctx(mock_client)
+
+        result = runner.invoke(app, ["task", "mine", "--workspace-id", "W1", "--limit", "3"])
+        assert result.exit_code == 0
+        data = json.loads(result.stdout)
+        assert data["truncated"] is True
+        assert data["count"] == 3
+
+
+# =============================================================================
+# Help text assertions (from polish-63)
+# =============================================================================
+
+
+class TestHelpText:
+    def test_list_stats_help_mentions_space_id(self) -> None:
+        from .conftest import _plain_help
+
+        result = runner.invoke(app, ["list", "stats", "--help"])
+        assert result.exit_code == 0
+        assert "use --space-id to narrow" in _plain_help(result)
+
+    def test_brief_help_is_compact(self) -> None:
+        from .conftest import _plain_help
+
+        result = runner.invoke(app, ["task", "list", "--help"])
+        assert result.exit_code == 0
+        assert "Compact projection" in _plain_help(result)
+        assert "date_updated" in _plain_help(result)
+
+    def test_brief_help_on_get(self) -> None:
+        from .conftest import _plain_help
+
+        result = runner.invoke(app, ["task", "get", "--help"])
+        assert result.exit_code == 0
+        assert "Compact projection" in _plain_help(result)
+
+    def test_brief_help_on_mine(self) -> None:
+        from .conftest import _plain_help
+
+        result = runner.invoke(app, ["task", "mine", "--help"])
+        assert result.exit_code == 0
+        assert "Compact projection" in _plain_help(result)
+
+    def test_brief_help_on_search(self) -> None:
+        from .conftest import _plain_help
+
+        result = runner.invoke(app, ["task", "search", "--help"])
+        assert result.exit_code == 0
+        assert "Compact projection" in _plain_help(result)
+
+    def test_brief_help_on_create(self) -> None:
+        from .conftest import _plain_help
+
+        result = runner.invoke(app, ["task", "create", "--help"])
+        assert result.exit_code == 0
+        assert "Compact projection" in _plain_help(result)
+
+    def test_brief_help_on_update(self) -> None:
+        from .conftest import _plain_help
+
+        result = runner.invoke(app, ["task", "update", "--help"])
+        assert result.exit_code == 0
+        assert "Compact projection" in _plain_help(result)
+
+    def test_brief_help_on_done(self) -> None:
+        from .conftest import _plain_help
+
+        result = runner.invoke(app, ["task", "done", "--help"])
+        assert result.exit_code == 0
+        assert "Compact projection" in _plain_help(result)
+
+    def test_brief_help_on_status(self) -> None:
+        from .conftest import _plain_help
+
+        result = runner.invoke(app, ["task", "status", "--help"])
+        assert result.exit_code == 0
+        assert "Compact projection" in _plain_help(result)
+
+    def test_bulk_export_help_mentions_default_list(self) -> None:
+        from .conftest import _plain_help
+
+        result = runner.invoke(app, ["bulk", "export-tasks", "--help"])
+        assert result.exit_code == 0
+        assert "Defaults to the configured default list" in _plain_help(result)
+
+    def test_task_group_help_mentions_comments(self) -> None:
+        from .conftest import _plain_help
+
+        result = runner.invoke(app, ["task", "--help"])
+        assert result.exit_code == 0
+        assert "comments" in _plain_help(result).lower()
+        assert "clickup task comments" in _plain_help(result)
+
+
+# =============================================================================
+# Comment.user optional (from polish-63)
+# =============================================================================
+
+
+class TestCommentUserOptional:
+    def test_comment_without_user_is_valid(self) -> None:
+        comment = Comment(id="c1", comment_text="hi", date="123")
+        assert comment.user is None
+
+    def test_comment_with_user_is_valid(self) -> None:
+        comment = Comment(
+            id="c1",
+            comment_text="hi",
+            date="123",
+            user=User(id=1, username="u", email="u@x.com"),
+        )
+        assert comment.user is not None
+        assert comment.user.username == "u"
